+++ lib/lake/utils.rb
@@ -1,0 +1,7 @@
+module Lake
+  module Utils
+    def lake_root
+      File.expand_path("../..",__FILE__)
+    end
+  end
+end