+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 Changelog
 ==
 
+v 0.1.5
+--
+* Added `lib/lake/utils.rb` with `Lake::Utils::lake_root`
+* Added support for lake-gen
+
+v 0.1.4
+--
+* Rewrote `need_build?`
+* Added `virtual` keyword
+
 v 0.1.3
 --
 * Fixed a mistake in `lake -T`